+++ cmd/api/routes.go
@@ -1,0 +1,22 @@
+package main
+
+import (
+	"github.com/go-chi/chi/v5"
+	"github.com/go-chi/chi/v5/middleware"
+	"net/http"
+)
+
+func (app *application) routes() http.Handler {
+	// Create a new Chi Router Mux
+	mux := chi.NewRouter()
+
+	// Define the routes
+	mux.Use(middleware.Recoverer)
+	mux.Use(app.enableCORS)
+
+	mux.Get("/", app.Home)
+
+	mux.Get("/movies", app.AllMovies)
+
+	return mux
+}